+++ src/main/java/cart/auth/AuthenticationInterceptor.java
@@ -1,7 +1,7 @@
 package cart.auth;
 
 import cart.auth.credential.CredentialDao;
-import cart.auth.credential.CredentialEntity;
+import cart.auth.credential.Credential;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
@@ -19,17 +19,14 @@ public class AuthenticationInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        CredentialEntity credential = basicAuthorizationExtractor.extract(request);
-
-        CredentialEntity savedCredential = credentialDao.findByEmail(credential.getEmail())
+        Credential credential = basicAuthorizationExtractor.extract(request);
+        Credential savedCredential = credentialDao.findByEmail(credential.getEmail())
                 .orElseThrow(()-> new IllegalArgumentException("존재하지 않는 회원입니다"));
-
         if(!credential.getPassword().equals(savedCredential.getPassword())) {
             throw new IllegalArgumentException("비밀번호가 일치하지 않습니다");
         }
-
         request.setAttribute("memberId", savedCredential.getMemberId());
-
         return true;
     }
+
 }
